// Angular2/play-ground/app/contacts/contact.interface.ts
/* Copyright (C) 2016 Pracxs Net & ITCE - All Rights Reserved
 * You may use, distribute and modify this code under the
 * terms of the Prometheus courses license.
 *
 * You should have received a copy of the Prometheus courses
 * license.If not, please write to: prometheus@pracxs.com
 * or to prometheus@itce.com
 */

interface Contact {
	id: number
	firstName: string
	lastName: string
	email?: string
}